+++ core/src/com/unciv/ui/screens/pickerscreens/PolicyPickerScreen.kt
@@ -17,7 +17,7 @@ import com.unciv.models.ruleset.PolicyBranch
 import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.models.translations.fillPlaceholders
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.BorderedTable
+import com.unciv.ui.components.widgets.BorderedTable
 import com.unciv.ui.components.ColorMarkupLabel
 import com.unciv.ui.components.extensions.addSeparator
 import com.unciv.ui.components.extensions.center
