+++ core/src/com/unciv/ui/screens/pickerscreens/PromotionButton.kt
@@ -4,7 +4,7 @@ import com.badlogic.gdx.scenes.scene2d.Touchable
 import com.badlogic.gdx.scenes.scene2d.ui.Label
 import com.badlogic.gdx.utils.Align
 import com.unciv.models.ruleset.unit.Promotion
-import com.unciv.ui.components.BorderedTable
+import com.unciv.ui.components.widgets.BorderedTable
 import com.unciv.ui.components.extensions.toLabel
 import com.unciv.ui.images.ImageGetter
 import com.unciv.ui.screens.basescreen.BaseScreen
